REF: Move test server files into separate directory from tests
