+++ manage.py
@@ -8,6 +8,7 @@ import sys
 
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.test_server.settings")
     from django.core.management import execute_from_command_line
+
     execute_from_command_line(sys.argv)
