+++ tests/test_server/apps.py
@@ -3,6 +3,6 @@ from django.utils.translation import gettext_lazy as _
 
 
 class DjangoTwinedTestAppConfig(AppConfig):
-    name = 'tests'
-    label = 'tests'
-    verbose_name = _('Django Twined Test App')
+    name = "tests"
+    label = "tests"
+    verbose_name = _("Django Twined Test App")
