+++ tests/test_server/urls.py
@@ -6,5 +6,5 @@ admin.autodiscover()
 
 
 urlpatterns = [
-    url(r'^admin/', admin.site.urls),
+    url(r"^admin/", admin.site.urls),
 ]
