+++ tox.ini
@@ -3,7 +3,7 @@ envlist = {py37,py38,py39}-dj{31}-channels{30}-{sqlite,postgres}
 
 [testenv]
 setenv =
-    DJANGO_SETTINGS_MODULE=tests.settings
+    DJANGO_SETTINGS_MODULE=tests.test_server.settings
     PYTHONPATH = {toxinidir}:{toxinidir}/django_twined
     sqlite: DATABASE_ENGINE=sqlite
     postgres: DATABASE_ENGINE=postgres
